requirements.txt
pandas
numpy
psycopg2-binary
apache-airflow[postgres]